cli/queries: split out deployment filters
